fix(proxyproto): removed unused code
